BackEnd/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.6.0
